python-sessions/requirements.txt
bcrypt==4.0.0
Django~=4.0.5
GitPython~=3.1.27
requests~=2.28.1